split the runstatus
