+++ photon_stream/production/__init__.py
@@ -1,6 +1,8 @@
 from . import tools
 from . import prepare
 from . import runinfo
-from . import status
+from . import runstatus
 from . import ethz
 from . import isdc
+from . import drs_run
+from . import show_runstatus
